ai-service/requirements.txt
flask==3.0.0
python-dotenv==1.0.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
google-generativeai==0.3.0
requests==2.31.0
numpy==1.24.3
torch==2.0.1
